separated Step Content to its own component
