+++ src/components/tutorialEditor/index.tsx
@@ -3,12 +3,9 @@ import styled, { AnyStyledComponent } from "styled-components";
 import "../app/normalize.css";
 
 import AceEditor from "react-ace";
-import { Icon } from "@iconify/react";
-import chevronUp from "@iconify-icons/feather/chevron-up";
-import chevronDown from "@iconify-icons/feather/chevron-down";
-import bxTrashAlt from "@iconify-icons/bx/bx-trash-alt";
 import { Navbar } from "../navbar";
 import { HintInput } from "./hintInput";
+import { StepContent } from "./stepContent";
 
 import "ace-builds";
 import "ace-builds/webpack-resolver";
@@ -18,10 +15,6 @@ import "ace-builds/src-noconflict/theme-tomorrow_night_eighties";
 
 import { background, grey } from "../../constants";
 import { Button } from "../textEditor/button";
-
-interface DeleteProps {
-  visible: string;
-}
 
 interface ContentBlock {
   type: string;
@@ -136,57 +129,6 @@ const ContentBlock: AnyStyledComponent = styled.div`
   margin-bottom: 20px;
 `;
 
-const ContentBlockInnerWrapper: AnyStyledComponent = styled.div`
-  padding: 20px;
-  width: ${(props: DeleteProps): string =>
-    props.visible === "flex" ? "93%" : "100%"};
-`;
-
-const UpDownContainer: AnyStyledComponent = styled.div`
-  position: relative;
-  float: right;
-  display: inline-block;
-  margin-bottom: 15px;
-`;
-
-const Arrow: AnyStyledComponent = styled(Icon)`
-  width: 40px;
-  height: 40px;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const TextInput: AnyStyledComponent = styled.textarea`
-  width: 100%;
-  min-height: 82px;
-  background-color: #46464e;
-  outline: none;
-  border: none;
-  border-radius: 1px;
-  padding: 5px;
-  color: #f5f5f5;
-  resize: none;
-`;
-
-const DeleteBlock: AnyStyledComponent = styled.div`
-  display: ${(props: DeleteProps): string => props.visible};
-  border-left: solid 1px #727272;
-  width: 7%;
-  text-align: center;
-  justify-content: center;
-  &:hover {
-    cursor: pointer;
-    background: rgba(255, 255, 255, 0.1);
-  }
-`;
-
-const TrashIcon: AnyStyledComponent = styled(Icon)`
-  align-self: center;
-  width: 22px;
-  height: 22px;
-`;
-
 const TutEditor: React.FC = (): JSX.Element => {
   const [content, setContent] = useState([
     {
@@ -195,55 +137,12 @@ const TutEditor: React.FC = (): JSX.Element => {
     }
   ]);
 
-  const [hovering, setHovering] = useState(-1);
-
-  const changeType = (type: string, index: number): void => {
-    const contentCopy = [...content];
-    contentCopy[index].type = type;
-    setContent(contentCopy);
-  };
-
-  const changeOrder = (direction: "up" | "down", index: number): void => {
-    const contentCopy = [...content];
-    if (direction === "up" && index - 1 >= 0) {
-      const temp = contentCopy[index - 1];
-      contentCopy[index - 1] = contentCopy[index];
-      contentCopy[index] = temp;
-    } else if (direction === "down" && index + 1 < content.length) {
-      const temp = contentCopy[index + 1];
-      contentCopy[index + 1] = contentCopy[index];
-      contentCopy[index] = temp;
-    }
-    setContent(contentCopy);
-  };
-
   const addBlock = (): void => {
     const contentCopy = [...content];
     contentCopy.push({
       type: "text",
       value: ""
     });
-    setContent(contentCopy);
-  };
-
-  const deleteBlock = (index: number): void => {
-    const contentCopy = [...content];
-    contentCopy.splice(index, 1);
-    setContent(contentCopy);
-  };
-
-  const handleTextChange = (
-    index: number,
-    event: React.ChangeEvent<HTMLInputElement>
-  ): void => {
-    const contentCopy = [...content];
-    contentCopy[index].value = event.target.value;
-    setContent(contentCopy);
-  };
-
-  const handleCodeChange = (index: number, code: string): void => {
-    const contentCopy = [...content];
-    contentCopy[index].value = code;
     setContent(contentCopy);
   };
 
@@ -281,102 +180,13 @@ const TutEditor: React.FC = (): JSX.Element => {
               <div>
                 {content.map((value: ContentBlock, index: number) => {
                   return (
-                    <ContentBlock
+                    <StepContent
+                      value={value}
+                      index={index}
                       key={index.toString()}
-                      onMouseEnter={(): void => {
-                        setHovering(index);
-                      }}
-                      onMouseLeave={(): void => {
-                        setHovering(-1);
-                      }}
-                    >
-                      <ContentBlockInnerWrapper>
-                        <Button
-                          backgroundColor={
-                            value.type === "text"
-                              ? "rgba(256, 256, 256, 0.2)"
-                              : "none"
-                          }
-                          submitFunction={(): void => {
-                            changeType("text", index);
-                          }}
-                          text="Text"
-                        />
-                        <Button
-                          backgroundColor={
-                            value.type === "code"
-                              ? "rgba(256, 256, 256, 0.2)"
-                              : "none"
-                          }
-                          submitFunction={(): void => {
-                            changeType("code", index);
-                          }}
-                          text="Code"
-                        />
-                        <UpDownContainer>
-                          <Arrow
-                            icon={chevronUp}
-                            onClick={(): void => {
-                              changeOrder("up", index);
-                            }}
-                          />
-                          <Arrow
-                            icon={chevronDown}
-                            onClick={(): void => {
-                              changeOrder("down", index);
-                            }}
-                          />
-                        </UpDownContainer>
-                        <TextInput
-                          value={content[index].value}
-                          style={{
-                            display:
-                              content[index].type === "text" ? "block" : "none"
-                          }}
-                          type="text"
-                          placeholder="Type text here..."
-                          onChange={(
-                            event: React.ChangeEvent<HTMLInputElement>
-                          ): void => {
-                            handleTextChange(index, event);
-                          }}
-                        />
-                        <AceEditor
-                          value={content[index].value}
-                          style={{
-                            position: "relative",
-                            marginTop: "0px",
-                            height: "82px",
-                            width: "100%",
-                            backgroundColor: background,
-                            fontFamily: "Source Code Pro",
-                            display:
-                              content[index].type === "code" ? "block" : "none"
-                          }}
-                          fontSize="16px"
-                          mode="python"
-                          theme="tomorrow_night_eighties"
-                          placeholder="Type code here..."
-                          setOptions={{
-                            enableBasicAutocompletion: true,
-                            enableLiveAutocompletion: true,
-                            enableSnippets: true,
-                            tabSize: 4
-                          }}
-                          onChange={(code: string): void => {
-                            handleCodeChange(index, code);
-                          }}
-                        />
-                      </ContentBlockInnerWrapper>
-                      <DeleteBlock
-                        visible={hovering === index ? "flex" : "none"}
-                        onClick={(): void => {
-                          deleteBlock(index);
-                        }}
-                      >
-                        <TrashIcon icon={bxTrashAlt} />
-                      </DeleteBlock>
-                    </ContentBlock>
+                      content={content}
+                      setContent={setContent}
+                    />
                   );
                 })}
               </div>
